Move CompositorLoader to satpy.composites.config_loader
