+++ satpy/scene.py
@@ -20,7 +20,8 @@
 import logging
 import os
 
-from satpy.composites import CompositorLoader, IncompatibleAreas
+from satpy.composites import IncompatibleAreas
+from satpy.composites.config_loader import CompositorLoader
 from satpy.config import get_environ_config_dir
 from satpy.dataset import (DataQuery, DataID, dataset_walker,
                            replace_anc, combine_metadata)
