+++ satpy/tests/compositor_tests/test_abi.py
@@ -25,7 +25,7 @@ class TestABIComposites(unittest.TestCase):
 
     def test_load_composite_yaml(self):
         """Test loading the yaml for this sensor."""
-        from satpy.composites import CompositorLoader
+        from satpy.composites.config_loader import CompositorLoader
         cl = CompositorLoader()
         cl.load_sensor_composites('abi')
